# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 스텝업(Step-Up) API
  version: 1.0.0
  description: |
    맞춤형 취업 로드맵 설계 서비스 '스텝업(Step-Up)' REST API 명세서
    
    ## 개요
    사용자의 현재 스펙을 분석하고, 실제 채용 공고 기반의 맞춤형 목표를 설정하여,
    달성을 위한 구체적인 로드맵을 제공하는 취업 준비 관리 서비스입니다.
    
    ## 인증
    간단한 구현을 위해 JWT 토큰 대신 `x-user-id` 헤더를 사용합니다.
    모든 API 요청 시 `x-user-id` 헤더에 사용자 ID를 포함해야 합니다.
    
  contact:
    name: Step-Up API Support
    email: support@stepup.com
servers:
  - url: http://localhost:8000
    description: 로컬 개발 서버
...
paths:
...
  /auth/register:
    post:
      tags: [인증]
      summary: 회원가입
      description: 새로운 사용자 계정을 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - password
                - email
              properties:
                user_id:
                  type: string
                  description: 사용자 아이디
                  example: "kim_frontend"
                password:
                  type: string
                  format: password
                  description: 비밀번호 (영문, 숫자, 특수문자 포함 8자 이상)
                  example: "Password123!"
                email:
                  type: string
                  format: email
                  description: 이메일 주소
                  example: "kim@example.com"
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "회원가입이 완료되었습니다"
                  user_id:
                    type: string
                    example: "kim_frontend"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: 중복된 아이디 또는 이메일
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [인증]
      summary: 로그인
      description: 아이디와 비밀번호로 로그인합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - password
              properties:
                user_id:
                  type: string
                  example: "kim_frontend"
                password:
                  type: string
                  format: password
                  example: "Password123!"
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "로그인 성공"
                  user_id:
                    type: string
                    example: "kim_frontend"
                  onboarding_completed:
                    type: boolean
                    description: 온보딩 완료 여부
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ========================================
  # 사용자 API
  # ========================================
...
components:
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: 오류 메시지
          example: "요청 처리 중 오류가 발생했습니다"
        code:
          type: string
          description: 오류 코드
          example: "INVALID_REQUEST"
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "필수 필드가 누락되었습니다"
            code: "BAD_REQUEST"
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "인증이 필요합니다"
            code: "UNAUTHORIZED"
...